update flow cbl template for new material distinction
