+++ template_project_flow/cbl/spruce1_0mm_cube1/spruce1_0mm_cube1-input.cwPar
@@ -1,0 +1,67 @@
+
+        // ================================================================================
+        // CHRONO CBL WORKBENCH - github.com/Cusatis-Computational-Services/CBL-preprocesser
+        //
+        // Copyright (c) 2025 
+        // All rights reserved. 
+        //
+        // Use of the code that generated this file is governed by a BSD-style license that
+        // can be found in the LICENSE file at the top level of the distribution and at
+        // github.com/Cusatis-Computational-Services/CBL-preprocesser/LICENSE
+        //
+        // ================================================================================
+        // CBL Workbench Parameter File
+        // ================================================================================
+        // 
+        // CBL Workbench developed by Cusatis Computational Services, Inc.
+        //
+        // ================================================================================
+        
+
+START
+current_time= 2025-07-14 12:21:52
+geoName=spruce1_0mm_cube1
+radial_growth_rule=binary_lloyd
+species=norwayspruce
+width_heart=0.15
+ring_width=2.0
+late_ratio=0.3
+cellsize_early=0.032
+cellsize_late=0.017
+cellwallthickness_early=0.0035
+cellwallthickness_late=0.011
+cell_length=3.0
+randomFlag=On
+dist_type=['TruncatedGaussian']
+dist_params=[[1.0, 0.01, 0.0]]
+corr_l=0.3
+sampling_type=MC
+box_shape=cube
+box_center=(15.75, 0.0)
+box_size=1.0
+box_height=1.0
+box_width=1.0000000001
+box_depth=1.0000000001
+x_notch_size=0
+y_notch_size=0
+precrackFlag=Off
+precrack_start=(5.5, 0)
+precrack_end=(5.65, 0)
+iter_max=2
+theta_min=0.0
+long_connector_ratio=0.05
+knotFlag=Off
+Uinf=1.0
+a1=0.4
+a2=0.1
+m1=0.05
+m2=0.05
+boundaryFlag=Off
+flowFlag=On
+mergeFlag=Off
+rayFlag=On
+inpType=all
+visFlag=On
+outDir=C:\Users\SusanAlexisBrown\woodWorkbench
+geoFile=
+
