--- template_project_flow/cbl_cube/spruce0_05mm_cube1.inp
@@ -1,218 +1,0 @@
-*Heading
-** Job name: prism_100mmx500mm Model name: Model prism_100mmx500mm-edgeEle
-*Preprint, echo=NO, model=NO, history=NO, contact=NO
-**
-** PARTS
-**
-*Part, name=Part-1
-*USER ELEMENT,NODES=2,TYPE=U1,PROPERTIES=8,COORDINATES=3,VARIABLES=4
-*NODE, NSET=NSET1
-1,-0.025,-0.025,0.0
-2,0.025,-0.025,0.0
-3,0.025,0.025,0.0
-4,-0.025,0.025,0.0
-5,0.001,0.015,0.0
-6,0.001,-0.01,0.0
-7,0.011,0.001,0.0
-8,-0.013,-0.001,0.0
-9,0.0,-0.025,0.0
-10,0.025,0.0,0.0
-11,0.0,0.025,0.0
-12,-0.025,0.0,0.0
-13,-0.025,-0.025,0.01219512
-14,0.025,-0.025,0.01219512
-15,0.025,0.025,0.01219512
-16,-0.025,0.025,0.01219512
-17,0.001,0.015,0.01219512
-18,0.001,-0.01,0.01219512
-19,0.011,0.001,0.01219512
-20,-0.013,-0.001,0.01219512
-21,0.0,-0.025,0.01219512
-22,0.025,0.0,0.01219512
-23,0.0,0.025,0.01219512
-24,-0.025,0.0,0.01219512
-25,-0.025,-0.025,0.025
-26,0.025,-0.025,0.025
-27,0.025,0.025,0.025
-28,-0.025,0.025,0.025
-29,0.001,0.015,0.025
-30,0.001,-0.01,0.025
-31,0.011,0.001,0.025
-32,-0.013,-0.001,0.025
-33,0.0,-0.025,0.025
-34,0.025,0.0,0.025
-35,0.0,0.025,0.025
-36,-0.025,0.0,0.025
-37,-0.025,-0.025,0.03780488
-38,0.025,-0.025,0.03780488
-39,0.025,0.025,0.03780488
-40,-0.025,0.025,0.03780488
-41,0.001,0.015,0.03780488
-42,0.001,-0.01,0.03780488
-43,0.011,0.001,0.03780488
-44,-0.013,-0.001,0.03780488
-45,0.0,-0.025,0.03780488
-46,0.025,0.0,0.03780488
-47,0.0,0.025,0.03780488
-48,-0.025,0.0,0.03780488
-49,-0.025,-0.025,0.05
-50,0.025,-0.025,0.05
-51,0.025,0.025,0.05
-52,-0.025,0.025,0.05
-53,0.001,0.015,0.05
-54,0.001,-0.01,0.05
-55,0.011,0.001,0.05
-56,-0.013,-0.001,0.05
-57,0.0,-0.025,0.05
-58,0.025,0.0,0.05
-59,0.0,0.025,0.05
-60,-0.025,0.0,0.05
-*ELEMENT, TYPE=U1, ELSET=UEL
-1,1,13
-2,2,14
-3,3,15
-4,4,16
-5,5,17
-6,6,18
-7,7,19
-8,8,20
-9,9,21
-10,10,22
-11,11,23
-12,12,24
-13,13,49
-14,14,50
-15,15,51
-16,16,52
-17,17,53
-18,18,54
-19,19,55
-20,20,56
-21,21,57
-22,22,58
-23,23,59
-24,24,60
-25,25,37
-26,26,38
-27,27,39
-28,28,40
-29,29,41
-30,30,42
-31,31,43
-32,32,44
-33,33,45
-34,34,46
-35,35,47
-36,36,48
-37,21,14
-38,21,13
-39,21,18
-40,23,16
-41,23,15
-42,23,17
-43,22,15
-44,22,14
-45,22,19
-46,15,17
-47,15,19
-48,14,18
-49,14,19
-50,20,18
-51,20,13
-52,20,24
-53,20,19
-54,20,17
-55,20,16
-56,18,19
-57,18,13
-58,13,24
-59,24,16
-60,19,17
-61,17,16
-62,45,38
-63,45,37
-64,45,42
-65,47,40
-66,47,39
-67,47,41
-68,46,39
-69,46,38
-70,46,43
-71,39,41
-72,39,43
-73,38,42
-74,38,43
-75,44,42
-76,44,37
-77,44,48
-78,44,43
-79,44,41
-80,44,40
-81,42,43
-82,42,37
-83,37,48
-84,48,40
-85,43,41
-86,41,40
-*UEL PROPERTY, ELSET=UEL
-1.000E-9,8.9E-10,2.15E+3,1.97E-14,6.1728E+4,0.5,0.0,17.28E+6
-*End Part
-**
-**
-** ASSEMBLY
-**
-*Assembly, name=Assembly
-**
-*Instance, name=Part-1-1, part=Part-1
-**Section: Section-1
-*Solid Section,elset=VISUALIZATION,material=Material-1
-*End Instance
-**
-*End Assembly
-*Amplitude, name=Amp-1, time=TOTAL TIME
-0., 0., 1.728e5, 1., 8.64e5, 1.
-**
-** MATERIALS
-**
-*Material, name=Material-1
-*Depvar
-4
-1, EPSILONV1, EPSILONV1
-2, EPSILONV2, EPSILONV2
-3, Vc1, Vc1
-4, Vc2, Vc2
-*Density
-0.,
-*User Material, constants=1, type=THERMAL
-0.,
-**
-** PREDEFINED FIELDS
-**
-** Name: Predefined Field-1   Type: Temperature
-*Initial Conditions, type=TEMPERATURE
-AllNodes, 0.0
-**
-** ----------------------------------------------------------------
-**
-** STEP: Step-1
-**
-*Step, name=Step-1, nlgeom=NO, inc=1000
-*Heat Transfer, end=PERIOD
-864, 8.64e5 
-**
-** BOUNDARY CONDITIONS
-**
-** Name: BC-1 Type: Temperature
-*Boundary, amp=Amp-1
-LeftNodes, 11, 11, 1.0
-**Boundary
-**RightNodes, 11, 11, 0.0
-**
-** OUTPUT REQUESTS
-**
-*Restart, write, frequency=0
-**
-** FIELD OUTPUT: F-Output-1
-**
-*Output, Field
-*End Step
