--- template_project_flow/cbl_cube/square_edgeEle.inp
@@ -1,108 +1,0 @@
-*Heading
-** Job name: Model square_edgeEle Model name: Model square_edgeEle
-*Preprint, echo=NO, model=NO, history=NO, contact=NO
-**
-** PARTS
-**
-*Part, name=Part-1
-*USER ELEMENT,NODES=2,TYPE=U1,PROPERTIES=8,COORDINATES=3,VARIABLES=4
-*NODE, NSET=NSET1
-1,-0.165216,-0.259751,0.0
-2,0.196350,0.031155,0.0
-3,-0.299116,0.191698,0.0
-4,-0.500000,-0.500000,0.0
-5,0.500000,-0.500000,0.0
-6,0.500000,0.500000,0.0
-7,-0.500000,0.500000,0.0
-8,0.000000,-0.500000,0.0
-9,0.500000,0.000000,0.0
-10,0.000000,0.500000,0.0
-11,-0.500000,0.000000,0.0
-12,-0.500000,0.250000,0.0
-*ELEMENT, TYPE=U1, ELSET=UEL
-1,9,6
-2,9,5
-3,9,2
-4,6,10
-5,6,2
-6,5,2
-7,5,8
-8,8,4
-9,8,2
-10,8,1
-11,10,7
-12,10,2
-13,10,3
-14,2,1
-15,2,3
-16,1,4
-17,1,11
-18,1,3
-19,12,11
-20,12,7
-21,12,3
-22,11,4
-23,11,3
-24,7,3
-*UEL PROPERTY, ELSET=UEL
-1.000E-9,8.9E-10,2.15E+3,1.97E-14,6.1728E+4,0.5,0.0,17.28E+6
-*End Part
-**
-**
-** ASSEMBLY
-**
-*Assembly, name=Assembly
-**
-*Instance, name=Part-1-1, part=Part-1
-**Section: Section-1
-*Solid Section,elset=VISUALIZATION,material=Material-1
-*End Instance
-**
-*End Assembly
-*Amplitude, name=Amp-1, time=TOTAL TIME
-0., 0., 1.728e5, 1., 8.64e5, 1.
-**
-** MATERIALS
-**
-*Material, name=Material-1
-*Depvar
-4
-1, EPSILONV1, EPSILONV1
-2, EPSILONV2, EPSILONV2
-3, Vc1, Vc1
-4, Vc2, Vc2
-*Density
-0.,
-*User Material, constants=1, type=THERMAL
-0.,
-**
-** PREDEFINED FIELDS
-**
-** Name: Predefined Field-1   Type: Temperature
-*Initial Conditions, type=TEMPERATURE
-AllNodes, 0.0
-**
-** ----------------------------------------------------------------
-**
-** STEP: Step-1
-**
-*Step, name=Step-1, nlgeom=NO, inc=1000
-*Heat Transfer, end=PERIOD
-864, 8.64e5 
-**
-** BOUNDARY CONDITIONS
-**
-** Name: BC-1 Type: Temperature
-*Boundary, amp=Amp-1
-LeftNodes, 11, 11, 1.0
-**Boundary
-**RightNodes, 11, 11, 0.0
-**
-** OUTPUT REQUESTS
-**
-*Restart, write, frequency=0
-**
-** FIELD OUTPUT: F-Output-1
-**
-*Output, Field
-*End Step
